requirements.txt
torch
torchaudio
rich
qwen-asr
transformers==4.57.6
accelerate
numpy
scipy
soundfile
openai
python-dotenv
onnxruntime-gpu>=1.16
huggingface_hub>=0.25
fastapi
uvicorn[standard]
pywebview
python-multipart
httpx
